fix: change lily58 labels to match other models
The lily58 is very close to the Sofle in layout but used LT for the
numbe row, and had the rows kinda offset by one. That is annoying
when you have multiple keyboards and want to use the same key position
for macros and combos.

This change makes the middle of the keyboard match so even if one uses
a smaller keyboard the alphas and numbers should match between them.

For people already using the current labels, they will have to change their
configs unfortunately, so it's a breaking change.