Harmonize lily58 key labels with standard convention (#84)
The lily58 key labels deviate from the convention set forth in github.com/urob/zmk-helpers#key-labels-collection. 

This commit harmonizes them with the usual convention. To retain backward-compatibility, the harmonized labels are guarded behind the `LILY58_STANDARD_LABELS` flag. To use the new harmonized labels, set the flag prior to loading the `lily58.h` header:

```c
#define LILY58_STANDARD_LABELS
#include <zmk-helpers/key-labels/lily58.h>
```

**Important**

A future commit will reverse the default. So if you are using the non-standard labels, you may want to revise your keymap to the standard labels, or watch out for breaking changes which will require setting a new flag to obtain the deprecated labels.